WIP: Proposal lecturer assignments
This PR proposes a schedule for who gives which lectures, with the intent that the other organizers will be on hand for other lectures to help run the room as problems occur. If you have any changes you'd like to suggest, please do so here before we merge in the schedule and make it live. Thanks.